agente-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
crewai==0.28.8
python-dotenv==1.0.0

